refactor(common.py): Move out network related functions to net.py
Also moves get_test_name to ci_tools.py
